Removes date settings from blog
