+++ app/brochure/views/developers/reference/blog.html
@@ -138,8 +138,7 @@
       <td><code>true</code> if you’ve indicated your avatar is round, <code>false</code> if not</td>
     </tr>
     <tr>
-      <td><code>\{{hideDates}}</code></td>
-      <td><code>true</code> if you’ve chosen to hide the dates on your site, <code>false</code> if not</td>
+     
     </tr>
   </tbody>
 </table>
