+++ app/brochure/views/templates/developers/reference/blog.html
@@ -137,9 +137,5 @@
       <td><code>\{{roundAvatar}}</code></td>
       <td><code>true</code> if you’ve indicated your avatar is round, <code>false</code> if not</td>
     </tr>
-    <tr>
-      <td><code>\{{hideDates}}</code></td>
-      <td><code>true</code> if you’ve chosen to hide the dates on your site, <code>false</code> if not</td>
-    </tr>
   </tbody>
 </table>
