+++ app/dashboard/routes/template-editor/index.js
@@ -32,7 +32,7 @@ TemplateEditor.route("/:templateSlug/settings")
   .all(require("./load/font-inputs"))
   .all(require("./load/color-inputs"))
   .all(require("./load/layout-inputs"))
-  .all(require("../settings/load/dates"))
+  .all(require("./load/dates"))
   .post(
     bodyParser,
     function (req, res, next) {
@@ -47,9 +47,6 @@ TemplateEditor.route("/:templateSlug/settings")
 
       req.locals = locals;
       req.partials = partials || {};
-
-      console.log(req.locals);
-      console.log(req.partials);
 
       next();
     },
