+++ app/dashboard/views/settings/services.html
@@ -3,8 +3,7 @@
 
   <a class="line" href="/settings/services/date">
     <span class="label">Date and time</span>
-    <span class="center">{{^blog.hideDates}}{{#displayFormats}}{{#selected}}{{date}}{{/selected}}{{/displayFormats}}{{/blog.hideDates}}
-      {{#blog.hideDates}}<i>Dates are hidden</i>{{/blog.hideDates}}</span>
+    <span class="center">{{currentTime}}</span>
   </a>
 
   <a class="line" href="/settings/services/permalinks">
