+++ app/models/blog/defaults.js
@@ -19,8 +19,6 @@ module.exports = {
   permalink: { format: "{{slug}}", custom: "", isCustom: false },
   timeZone: "UTC",
   dateFormat: "M/D/YYYY",
-  dateDisplay: "MMMM D, Y",
-  hideDates: false,
   forceSSL: true,
   redirectSubdomain: true,
   plugins: defaultPlugins,
